n3 circuit format
