+++ circuit/n3/multisig_verify_test.go
@@ -1,9 +1,7 @@
-package circuit
+package n3
 
 import (
 	native_crypto "crypto/ecdsa"
-	"crypto/elliptic"
-	"fmt"
 	"github.com/consensys/gnark-crypto/ecc"
 	"github.com/consensys/gnark/frontend"
 	"github.com/consensys/gnark/std/math/emulated"
@@ -155,53 +153,6 @@ func TestCheckPubKeyFormatCircuit(t *testing.T) {
 		err = test.IsSolved(&circuit, &witness, ecc.BN254.ScalarField())
 	}
 	assert.NoError(err)
-}
-
-func DecompressPubkey(pubkey []byte) (*native_crypto.PublicKey, error) {
-	x, y := new(big.Int), new(big.Int)
-	if len(pubkey) != 33 {
-		return nil, fmt.Errorf("invalid public key")
-	}
-	if (pubkey[0] != 0x02) && (pubkey[0] != 0x03) {
-		return nil, fmt.Errorf("invalid public key")
-	}
-	if x == nil {
-		return nil, fmt.Errorf("invalid public key")
-	}
-	x.SetBytes(pubkey[1:])
-
-	xxx := new(big.Int).Mul(x, x)
-	xxx.Mul(xxx, x)
-
-	ax := new(big.Int).Mul(big.NewInt(3), x)
-
-	yy := new(big.Int).Sub(xxx, ax)
-	yy.Add(yy, elliptic.P256().Params().B)
-
-	y1 := new(big.Int).ModSqrt(yy, elliptic.P256().Params().P)
-	if y1 == nil {
-		return nil, fmt.Errorf("can not revcovery public key")
-	}
-
-	y2 := new(big.Int).Neg(y1)
-	y2.Mod(y2, elliptic.P256().Params().P)
-
-	if pubkey[0] == 0x02 {
-		if y1.Bit(0) == 0 {
-			y = y1
-		} else {
-			y = y2
-		}
-	} else {
-		if y1.Bit(0) == 1 {
-			y = y1
-		} else {
-			y = y2
-		}
-	}
-	//fmt.Println("dx:",x)
-	//fmt.Println("dy:",y)
-	return &native_crypto.PublicKey{X: x, Y: y, Curve: elliptic.P256()}, nil
 }
 
 type TempForamtCheckStruct[T, S emulated.FieldParams] struct {
