+++ circuit/neox/keccak256_test.go
@@ -12,7 +12,7 @@ import (
 
 func TestKecc256Circuit(t *testing.T) {
 	assert := test.NewAssert(t)
-	header := new(EthBlockHeader)
+	header := new(NeoxBlockHeader)
 	err := header.UnmarshalJSON([]byte(
 		`{
 			"baseFeePerGas": "0x4a817c800",
